+++ resources/views/bootstrap-5/input.blade.php
@@ -27,7 +27,7 @@
                 'id' => $id,
                 'class' => 'form-control' . ($validationClass ? ' ' . $validationClass : null),
                 'type' => $type,
-                'name' => $isWired ? null : ($locale ? $name . '[' . $locale . ']' : $name),
+                'name' => $locale ? $name . '[' . $locale . ']' : $name,
                 'placeholder' => $placeholder,
                 'data-locale' => $locale,
                 'value' => $isWired ? null : ($value ?? ''),