+++ resources/views/bootstrap-4/input.blade.php
@@ -13,11 +13,11 @@
     @endphp
     <div @class(['d-none' => $type === 'hidden', 'form-floating' => $displayFloatingLabel, 'mb-3' => $marginBottom])>
         @if(($prepend || $append) && ! $displayFloatingLabel)
-            <x-form::partials.label :id="$id" class="$type === 'file' ? 'custom-file-label' : 'form-label'" :label="$label"/>
+            <x-form::partials.label :id="$id"  class="form-label" :label="$label"/>
             <div class="input-group">
         @endif
-            @if(! $prepend && ! $append && ! $displayFloatingLabel && $type !== 'file')
-                <x-form::partials.label :id="$id" :class="$type === 'file' ? 'custom-file-label' : 'form-label'" :label="$label"/>
+            @if(! $prepend && ! $append && ! $displayFloatingLabel)
+                <x-form::partials.label :id="$id" class="form-label" :label="$label"/>
             @endif
             @if($prepend && ! $displayFloatingLabel)
                 <div class="input-group-prepend">
@@ -39,7 +39,9 @@
                     'aria-describedby' => $caption ? $id . '-caption' : null,
                 ]) }}/>
             @if(! $prepend && ! $append && ($displayFloatingLabel || $type === 'file'))
-                <x-form::partials.label :id="$id" :class="$type === 'file' ? 'custom-file-label' : 'form-label'" :label="$label"/>
+                <x-form::partials.label :id="$id"
+                                        :class="$type === 'file' ? 'custom-file-label' : 'form-label'"
+                                        :label="$type === 'file' ? $placeholder : $label"/>
             @endif
             @if($type === 'file')
                 </div>