+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this package will be documented in this file.
 
 ## [1.0.2](https://github.com/Okipa/laravel-form-components/compare/1.0.1...1.0.2)
 
+2022-02-23
+
 * Fixed Bootstrap4 file input
 * Fixed Bootstrap4 caption partial
 
 ## [1.0.1](https://github.com/Okipa/laravel-form-components/compare/1.0.0...1.0.1)
+
+2022-02-23
 
 * Fixed Bootstrap4 toggle-switch classes